+++ dp2d.h
@@ -8,6 +8,6 @@
 struct fptree;
 
 void dp2d(const struct fptree *fp, double eps, double eps_share,
-		size_t k, double minalpha, long int seed);
+		size_t k, size_t minalpha, long int seed);
 
 #endif
